+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/PrintBundledModulesTask.kt
@@ -2,18 +2,21 @@
 
 package org.jetbrains.intellij.platform.gradle.tasks
 
-import com.jetbrains.plugin.structure.intellij.plugin.module.IdeModule
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
+import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.UntrackedTask
 import org.jetbrains.intellij.platform.gradle.Constants.Plugin
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
-import org.jetbrains.intellij.platform.gradle.services.IdesManagerService
+import org.jetbrains.intellij.platform.gradle.intellijPlatformIdeLayoutIndicesCachePath
+import org.jetbrains.intellij.platform.gradle.services.IdeLayoutIndexService
 import org.jetbrains.intellij.platform.gradle.services.registerClassLoaderScopedBuildService
 import org.jetbrains.intellij.platform.gradle.tasks.aware.IntelliJPlatformVersionAware
+import org.jetbrains.intellij.platform.gradle.utils.asPath
+import org.jetbrains.intellij.platform.gradle.utils.rootProjectPath
 
 /**
  * Prints the list of bundled modules available within the currently targeted IntelliJ Platform.
@@ -22,16 +25,24 @@ import org.jetbrains.intellij.platform.gradle.tasks.aware.IntelliJPlatformVersio
 abstract class PrintBundledModulesTask : DefaultTask(), IntelliJPlatformVersionAware {
 
     @get:Internal
-    abstract val idesManager: Property<IdesManagerService>
+    /**
+     * Shared service that provides the cached serialized IDE layout index for the target platform.
+     */
+    internal abstract val ideLayoutIndexService: Property<IdeLayoutIndexService>
+
+    @get:Internal
+    /**
+     * On-disk cache location for layout-index snapshots derived from extracted IDE distributions.
+     */
+    abstract val ideLayoutIndexCacheDirectory: DirectoryProperty
 
     @TaskAction
     fun printBundledModules() {
         println("Bundled modules for ${productInfo.name} ${productInfo.version} (${productInfo.buildNumber}):")
-        val ide = idesManager.get().resolve(platformPath)
+        val ideLayoutIndex = ideLayoutIndexService.get().resolve(platformPath, ideLayoutIndexCacheDirectory.asPath)
 
-        ide.bundledPlugins.asSequence()
-            .filterIsInstance<IdeModule>()
-            .map { it.pluginId }
+        ideLayoutIndex.bundledModules
+            .map { it.id }
             .toSet()
             .sorted()
             .forEach { println(it) }
@@ -45,7 +56,12 @@ abstract class PrintBundledModulesTask : DefaultTask(), IntelliJPlatformVersionA
     companion object : Registrable {
         override fun register(project: Project) =
             project.registerTask<PrintBundledModulesTask>(Tasks.PRINT_BUNDLED_MODULES) {
-                idesManager.convention(project.gradle.registerClassLoaderScopedBuildService(IdesManagerService::class))
+                ideLayoutIndexService.convention(project.gradle.registerClassLoaderScopedBuildService(IdeLayoutIndexService::class))
+                // Reuse the same cache location as dependency resolution so task output reflects the
+                // same indexed view of the extracted IDE layout.
+                ideLayoutIndexCacheDirectory.convention(project.layout.dir(project.provider {
+                    project.providers.intellijPlatformIdeLayoutIndicesCachePath(project.rootProjectPath).get().toFile()
+                }))
             }
     }
 }